Add files via upload
The system successfully recognizes and identifies the signature holder accurately with the
forgery issue gift in it. The popularity pattern is trained on Convolutional Neural Networks that
works well with the dataset of 1320 pictures and therefore the forgery detection is trained on the
whole image set of the individual that is around twenty-five pictures and every time the
calculations are runtime that minimizes the likelihood of error in classification.
A robust and reliable signature recognition and verification system with maximum
accuracy possible is very important for many purposes like enforcement, security management,
and lots of business processes. It can be used as an intermediate tool to authenticate several
documents like cheques, legal records, certificates, etc.The model gave encouraging results.
Entirely different threshold values are used for feature matching on testing and training vectors,
which helped to boost the overall performance and efficiency of the system.